Add flickr oauth token volumes
In order to write with the Flickr API you need more than just the key
and secret. You also need to generate the file oauth-tokens.sqlite and
put it in the .flickr directory in the home directory of the user
running the process. In this case it is /root/.flickr . The pysh project
has a script for generating this token (getflickrtoken.py).