Add health check
This adds a view at /health that displays the build commit hash
and the number of results in the database. Docker can use this to
determine basic health of the container. During an upgrade if the
new container does not pass the health check the previous container
will continue to be used.